requirements.txt
colorlog==2.10.0
Flask==0.12
Flask-Compress==1.4.0
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-WTF==0.14.2
gevent==1.2.1
gevent-socketio==0.3.6
gevent-websocket==0.9.5
greenlet==0.4.12
pymongo==3.4.0
requests==2.13.0
Werkzeug==0.11.15
zeep==1.4.1
git+https://github.com/kulikjak/skautis-py.git#egg=skautis-py
